move main.go logic out of commands
